requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
msgpack==1.0.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
